Addon Feature: Added colored text support
Use a simple and dirty escape sequence to set the color for the next
sequence of text. The sequence is "\x1b" followed by two hex digits.
"\x1b" is the escape character. The first hex digit denotes the
foreground color, and the second denotes the background color. The
default setting is "\x1b70" (white text on black background).